% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_old_name.R
\name{find_old_name}
\alias{find_old_name}
\title{Generate the a random set of 10 suggested neutral(by default) baby names based on the given time period and sex.}
\usage{
find_old_name(tp, limit = 10, sex = "uni")
}
\arguments{
\item{tp}{A character vector that specifying time period from 1880 to 2018}

\item{limit}{A double vector that is the number of name in the output list. Default 10}

\item{sex}{A character vector that is the aimed sex of the name in the output. Default "uni"}
}
\value{
A list contains random suggested old names based on the given limitation.
}
\description{
Generate the a random set of 10 suggested neutral(by default) baby names based on the given time period and sex.
}
\examples{
find_old_name('80s')
}
