% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar_name.R
\name{find_similar_name}
\alias{find_similar_name}
\title{Generate the a random set of 10 (by default) suggested similar baby names based on the syllable of the input name.}
\usage{
find_similar_name(name, limit = 10)
}
\arguments{
\item{name}{A character vector that is the given name used to find similar names based on syllable}

\item{limit}{A double vector that is the number of name in the output list. Default 10}
}
\value{
A list contains random suggested similar names based on the given name.
}
\description{
Generate the a random set of 10 (by default) suggested similar baby names based on the syllable of the input name.
}
\examples{
find_similar_name('Rowan')
}
