% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameformeR.R
\name{find_similar_name}
\alias{find_similar_name}
\title{Generate a random list of names that sound similar to a given user input name.
Uses Match rating approach algorithm to determine similarity}
\usage{
find_similar_name(match_name, limit = 10)
}
\arguments{
\item{match_name}{A character vector that is the name for comparison; output names will sound like this one}

\item{limit}{An integer vector that is the number of names in the output list. Default 10}
}
\value{
A vector containing random suggested similar names based on the given name.
}
\description{
Generate a random list of names that sound similar to a given user input name.
Uses Match rating approach algorithm to determine similarity
}
\examples{
find_similar_name('Elizabeth', 5)
}
