% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unisex_name.R
\name{find_unisex_name}
\alias{find_unisex_name}
\title{Generate the a random set of 10 suggested neutral baby names based on the given limitation and baby names in the past years.}
\usage{
find_unisex_name(limit = 0.2)
}
\arguments{
\item{limit}{A double vector that is the minimum proportion of the neutral names in a single year in the data frame. Default 0.2}
}
\value{
A list contains random 10 suggested neutral names based on the given limitation
}
\description{
Generate the a random set of 10 suggested neutral baby names based on the given limitation and baby names in the past years.
}
\examples{
find_unisex_name(0.0001)
}
